requirements.txt
Flask
Jinja2
MarkupSafe
Werkzeug
itsdangerous
python-dateutil
pytz
six
tzlocal
wsgiref
docker-py
requests[security]
github3.py
pytest
unittest2
mock
flake8
ansi2html
cherrypy
numpy
wheel
pyparsing
ipython
ndg-httpsclient
PyYaml
